feat(grammar): migrate to tree-sitter 0.26 + tree-sitter-kotlin ABI-15 (#268)
* feat(grammar): migrate to tree-sitter 0.26 + tree-sitter-kotlin ABI-15

The canonical tree-sitter-kotlin crate has been stuck at 0.3.8 (ABI 14) on
crates.io for ~2 years, with an open unanswered "please cut a release"
issue, while upstream (fwcd/tree-sitter-kotlin main) kept receiving real
grammar fixes. Bumps core tree-sitter 0.22 -> 0.26, swaps to BrokkAI's
0.4.0 fork (tracks upstream main, aliased under the same crate name so no
call sites needed renaming), and bumps tree-sitter-java 0.21 -> 0.23.

Directly confirmed fixed: the constructor-on-newline idiom (the common
Android @JvmOverloads View-constructor pattern) and `fun interface` with a
body -- both previously produced parse errors, both now parse cleanly.

tree-sitter-swift-bundled's published crate hard-pins tree-sitter = "0.22",
which conflicts with the core bump (Cargo's `links` uniqueness rule) even
though the vendored grammar itself doesn't care which binding version loads
it. Vendored locally (vendor/tree-sitter-swift-bundled) with the version
bound relaxed -- same compiled grammar, nothing else changed.

Fallout from the grammar bump, all fixed:
- Node::child/named_child take u32 now, not usize (mechanical, ~15 sites).
- QueryCursor::matches() returns a StreamingIterator, not Iterator.
- language() fns replaced by LANGUAGE consts (LanguageFn) needing .into().
- Kotlin extension fun/val receivers moved one level deeper, wrapped in a
  new `receiver_type` node -- extract_extension_receiver_from_cst updated.
- `null` became a named `null_literal` node (was an anonymous token).
- A real tree-sitter core regression for the vendored Swift grammar (ABI 13,
  external scanner): both descendant_for_point_range and
  descendant_for_byte_range silently stop at a coarse ancestor for any
  sibling after the first at a nesting level, even though start_byte/
  end_byte read correctly off the same nodes throughout. Worked around with
  a manual byte-containment walk seeded from the built-in's own result (a
  no-op for Kotlin/Java, which aren't affected).
- The `constructor` keyword no longer has any node in the new grammar at
  all (not renamed -- genuinely gone); recovered its semantic-highlighting
  token via a manual text scan within primary_constructor's span.
- New, narrow grammar ambiguity: a single-line `object Name { ... }` body
  misparses as an infix_expression (has_error() stays false). Confirmed not
  newline-sensitive is wrong -- multi-line (the idiomatic style) is
  unaffected; adjusted the 3 affected test fixtures accordingly.

Verified: 1760/1760 tests pass, clippy clean, fmt clean. Full
resolution-accuracy benchmark run against real nowInAndroid (338 files)
and a partial run against a ~13,200-file production monorepo (Moneta) --
no crashes, no hangs, steady throughput throughout.


* fix(grammar): repair extension-receiver resolution and drop dead ERROR-recovery code

The missing-import diagnostic had its own copy of extension-receiver
extraction that wasn't updated for the ABI-15 grammar's receiver_type
wrapper node, so it always failed to find the receiver — every
implicit-receiver call inside an extension function/property (Compose's
`item {}` DSL, androidx's `requireActivity()`, etc.) was wrongly flagged
as a missing import. Confirmed via a real production monorepo: 264 -> 28
false-positive flags after the fix, matching the pre-migration baseline.
Consolidated the two copies into one shared `extension_receiver_from_decl`
in parser.rs so they can't desync again, and added regression tests that
verify against the unfixed code before trusting them.

Also removed three ERROR-recovery workarounds in parser.rs that the new
grammar makes unreachable (fun-interface misparse recovery, chained-call
assignment suppression, nullable-function-type suppression) — verified
via the existing test suite plus a real-world syntax-error parity check
against the same monorepo (253 errors / 72 files, identical before and
after). A fourth workaround (file-annotation comma) turned out to still
be needed; kept it after its regression test failed without it.


* fix(grammar): address Copilot review findings on the point-lookup and constructor-token helpers

descendant_for_point: use saturating_add for offset + point.column so an
out-of-range caller position can't wrap on 32-bit usize instead of just
clamping to the line end as intended.

kotlin_primary_constructor_keyword_token: replace the raw substring search
for "constructor" with a trivia-skip (whitespace + // and /* */ comments)
to the next real token. The substring search could match the word inside
a comment sitting in the modifiers-to-'(' gap and highlight the wrong
range as the keyword. Added a regression test that fails against the old
substring search and passes with the fix.


* fix(ci): silence clippy::map_or_identity in trivia_skipped_offset

CI runs clippy with --all-targets, which the local pre-commit hook's
plain `cargo clippy` doesn't cover — .map_or(default, |x| x) is just
.unwrap_or(default).


---------

+++ src/indexer/infer/chain.rs
@@ -921,7 +921,7 @@ fn infer_lambda_result_type(
     if count == 0 {
         return None;
     }
-    let last = statements.named_child(count - 1)?;
+    let last = statements.named_child(count as u32 - 1)?;
     resolve_root_node_type(last, bytes, deps, uri)
 }
